# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductComment } from "./productComment";

export interface Product {
  id: string;
  name: string;
  description?: string;
  price: number;
  imageUrl: string;
  category: string;
  ownerId: string;
  ownerName: string;
  likesCount: number;
  userLiked: boolean;
  comments: ProductComment[];
  createdAt: string;
}