# lib/api-zod/src/generated/types/updateProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProductBody {
  userId: string;
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minimum 0 */
  price: number;
}